Consolidate tests and benchmarks smoke coverage
Merge redundant unittest modules into focused contract tests:
- test_catalog_plant_contracts (CartPole, DoublePendulum, DynamicBicycle)
- test_benchmark_smoke (f-evaluator, trajopt, optimizer, simulation)
- planning_helpers shared fixture for RRT/RRT*

Remove duplicate camera-follow table from catalog migration (covered by
test_camera_transform and representative plant checks). Drop trivial
metric constants test, commented JAX duplicate block in compile tests,
and benchmark-internal accounting/hygiene tests.

Update tests/README.md and benchmarks/README.md with suite philosophy.
