+++ tests/unittest/test_overlays.py
@@ -19,7 +19,11 @@ from minilink.graphical.animation.drawables import (
     SceneHistory,
     validate_overlay,
 )
-from minilink.graphical.animation.primitives import CustomLine, HorizonPolyline
+from minilink.graphical.animation.primitives import (
+    CustomLine,
+    HorizonPolyline,
+    TrajectoryPolyline,
+)
 from minilink.graphical.catalog import SceneHistory as SceneHistoryExport
 from minilink.planning.spatial.scene import Scene
 
@@ -83,6 +87,29 @@ class TestReplay(unittest.TestCase):
         ghost = Replay(plant, traj)
         ghost.tf(t=0.5)
         self.assertAlmostEqual(plant.last_t, 0.5)
+
+
+class TestTrajectoryPolyline(unittest.TestCase):
+    def test_prefix_window_grows_with_playback_time(self):
+        traj = Trajectory(
+            t=np.array([0.0, 1.0, 2.0, 3.0]),
+            x=np.array(
+                [
+                    [0.0, 1.0, 2.0, 3.0],
+                    [0.0, 0.2, 0.1, 0.0],
+                    np.zeros(4),
+                    np.full(4, 5.0),
+                    np.zeros(4),
+                    np.zeros(4),
+                ]
+            ),
+            u=np.zeros((2, 4)),
+        )
+        prim = TrajectoryPolyline(traj, window="prefix")
+        pts_early = prim.compute_pts(1.0)
+        pts_late = prim.compute_pts(2.5)
+        self.assertEqual(pts_early.shape[0], 2)
+        self.assertEqual(pts_late.shape[0], 3)
 
 
 class TestAnimatorOverlays(unittest.TestCase):
